Data Annotations & DTOs

+++ Back/src/ProEventos.API/Controllers/EventosController.cs
@@ -1,14 +1,11 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.Extensions.Logging;
 using ProEventos.Application.Contratos;
+using ProEventos.Application.Dtos;
 using ProEventos.Domain;
-using ProEventos.Persistence;
-using ProEventos.Persistence.Contextos;
+
 
 namespace ProEventos.API.Controllers
 {
@@ -22,7 +19,7 @@ namespace ProEventos.API.Controllers
         {
             _eventoService = eventoService;
 
-        }       
+        }
 
         [HttpGet]
         public async Task<IActionResult> Get()
@@ -30,17 +27,18 @@ namespace ProEventos.API.Controllers
             try
             {
                 var eventos = await _eventoService.GetAllEventosAsync(true);
-                if (eventos == null) return NotFound("Nenhum evento encontrado");
+                if (eventos == null) return NoContent();
+
                 
 
                 return Ok(eventos);
             }
             catch (Exception ex)
             {
-                
+
                 return this.StatusCode(StatusCodes.Status500InternalServerError, $"Erro ao tentar recuperar eventos. Erro {ex.Message}");
             }
-            
+
         }
 
         [HttpGet("{id}")]
@@ -49,14 +47,14 @@ namespace ProEventos.API.Controllers
             try
             {
                 var evento = await _eventoService.GetEventoByIdAsync(id, true);
-                if (evento == null) return NotFound("Evento por id não encontrado.");
-                
+                if (evento == null) return NoContent();
+
 
                 return Ok(evento);
             }
             catch (Exception ex)
             {
-                
+
                 return this.StatusCode(StatusCodes.Status500InternalServerError, $"Erro ao tentar recuperar evento. Erro {ex.Message}");
             }
         }
@@ -67,50 +65,50 @@ namespace ProEventos.API.Controllers
             try
             {
                 var evento = await _eventoService.GetAllEventosByTemaAsync(tema, true);
-                if (evento == null) return NotFound("Evento por tema não encontrado");
-                
+                if (evento == null) return NoContent();
+
 
                 return Ok(evento);
             }
             catch (Exception ex)
             {
-                
+
                 return this.StatusCode(StatusCodes.Status500InternalServerError, $"Erro ao tentar recuperar evento. Erro {ex.Message}");
             }
         }
 
         [HttpPost]
-        public async Task<IActionResult> Post(Evento model)
+        public async Task<IActionResult> Post(EventoDto model)
         {
             try
             {
                 var evento = await _eventoService.AddEventos(model);
-                if (evento == null) return BadRequest("Erro ao tentar adicionar evento");
-                
+                if (evento == null) return NoContent();
+
 
                 return Ok(evento);
             }
             catch (Exception ex)
             {
-                
+
                 return this.StatusCode(StatusCodes.Status500InternalServerError, $"Erro ao tentar adicionar evento. Erro {ex.Message}");
             }
         }
 
         [HttpPut("{id}")]
-        public async Task<IActionResult> Put(int id, Evento model)
+        public async Task<IActionResult> Put(int id, EventoDto model)
         {
             try
             {
                 var evento = await _eventoService.UpdateEvento(id, model);
-                if (evento == null) return BadRequest("Erro ao tentar atualizar evento");
-                
+                if (evento == null) return NoContent();
+
 
                 return Ok(evento);
             }
             catch (Exception ex)
             {
-                
+
                 return this.StatusCode(StatusCodes.Status500InternalServerError, $"Erro ao tentar recuperar evento. Erro {ex.Message}");
             }
         }
@@ -120,14 +118,18 @@ namespace ProEventos.API.Controllers
         {
             try
             {
-                return await _eventoService.DeleteEvento(id) ? 
-                    Ok("Evento deletado.") : 
-                    BadRequest("Evento não deletado");
-                
+
+                var evento = await _eventoService.GetEventoByIdAsync(id, true);
+                if (evento == null) return NoContent();
+
+                return await _eventoService.DeleteEvento(id) ?
+                    Ok("Evento deletado.") :
+                    throw new Exception("Ocorreu um problema não especificado ao tentar deletar o evento");
+
             }
             catch (Exception ex)
             {
-                
+
                 return this.StatusCode(StatusCodes.Status500InternalServerError, $"Erro ao tentar deletar eventos. Erro {ex.Message}");
             }
         }
